+++ src/backend/joanie/tests/signature/test_backend_signature_base.py
@@ -74,11 +74,12 @@ class BaseSignatureBackendTestCase(TestCase):
     @override_settings(
         JOANIE_SIGNATURE_BACKEND="joanie.signature.backends.dummy.DummySignatureBackend"
     )
-    def test_backend_signature_base_backend_confirm_student_signature(self):
+    def test_backend_signature_base_backend_confirm_signature(self):
         """
-        This test verifies that the `confirm_student_signature` method updates the contract with a
-        timestamps for the field 'student_signed_on', and it should not set 'None' to the field
-        'submitted_for_signature_on'.
+        This test verifies that the `confirm_signature` method updates the contract with a
+        timestamps for the field `student_signed_on` or `organization_signed_on` according to
+        the contract state and it should set to `None` the field `submitted_for_signature_on` if
+        the contract is signed by the organization.
 
         Furthermore, it should update the order state.
         """
@@ -90,16 +91,22 @@ class BaseSignatureBackendTestCase(TestCase):
         backend = get_signature_backend()
 
         order.submit_for_signature(order.owner)
-        backend.confirm_student_signature(
-            reference=contract.signature_backend_reference
-        )
+        backend.confirm_signature(reference=contract.signature_backend_reference)
 
         contract.refresh_from_db()
         self.assertIsNotNone(contract.submitted_for_signature_on)
         self.assertIsNotNone(contract.student_signed_on)
+        self.assertIsNone(contract.organization_signed_on)
 
         order.refresh_from_db()
         self.assertEqual(order.state, enums.ORDER_STATE_COMPLETED)
+
+        # Calling confirm_signature again should validate organization signature
+        backend.confirm_signature(reference=contract.signature_backend_reference)
+        contract.refresh_from_db()
+        self.assertIsNone(contract.submitted_for_signature_on)
+        self.assertIsNotNone(contract.student_signed_on)
+        self.assertIsNotNone(contract.organization_signed_on)
 
     @override_settings(
         JOANIE_SIGNATURE_BACKEND=random.choice(
@@ -110,11 +117,11 @@ class BaseSignatureBackendTestCase(TestCase):
         ),
         JOANIE_SIGNATURE_VALIDITY_PERIOD_IN_SECONDS=60 * 60 * 24 * 15,
     )
-    def test_backend_signature_base_backend_confirm_student_signature_but_validity_period_is_passed(
+    def test_backend_signature_base_backend_confirm_signature_but_validity_period_is_passed(
         self,
     ):
         """
-        This test verifies that the `confirm_student_signature` method does not update the contract
+        This test verifies that the `confirm_student` method does not update the contract
         if the validity period is passed. It should raise an error mentionning the validity of
         the signature is passed.
         """
@@ -134,7 +141,7 @@ class BaseSignatureBackendTestCase(TestCase):
         backend = get_signature_backend()
 
         with self.assertRaises(ValidationError) as context:
-            backend.confirm_student_signature(reference="wfl_fake_dummy_id")
+            backend.confirm_signature(reference="wfl_fake_dummy_id")
 
         self.assertEqual(
             str(context.exception),